Initial example states
