+++ game/utils.py
@@ -1,5 +1,6 @@
 import dataclasses
 import math
+from functools import partial
 from typing import ClassVar, Tuple
 
 
@@ -54,3 +55,4 @@ Vector2.UP = Vector2(0, -1)
 Vector2.DOWN = Vector2(0, 1)
 Vector2.LEFT = Vector2(-1, 0)
 Vector2.RIGHT = Vector2(1, 0)
+echo = partial(print, end='', flush=True)
